desktop and ios libs with new app setup
