+++ deps/window/src/lib.rs
@@ -1,6 +1,6 @@
-mod app;
 mod text;
 mod window;
+mod window_events;
 
 mod frame_counter;
 pub mod image;
@@ -9,7 +9,6 @@ mod state;
 mod surface;
 mod vertex_buffer;
 
-pub use app::*;
 pub use bytemuck::cast_slice;
 pub use image_proc::include_images;
 pub use screenshot::*;
@@ -21,6 +20,7 @@ pub use wgpu::{
     util::{BufferInitDescriptor, DeviceExt},
 };
 pub use window::*;
+pub use window_events::*;
 pub use winit::{
     event::{ElementState, MouseButton},
     keyboard::NamedKey,
