--- level-test/src/interface/mod.rs
@@ -1,1 +1,0 @@
-pub mod level_test_view;
