--- level-test/src/levels/mod.rs
@@ -1,3 +1,0 @@
-mod friction_level;
-
-pub use friction_level::*;
