+++ ui-test/src/base/layout.rs
@@ -1,6 +1,6 @@
 use anyhow::Result;
 use test_engine::{
-    App,
+    AppRunner,
     refs::Weak,
     ui::{
         Anchor, Button, Color, HasText, Setup, UI, ViewData, ViewSubviews,
@@ -102,7 +102,7 @@ impl Setup for LayoutPlace {
 pub async fn test_layout() -> Result<()> {
     UI::init_test_view::<LayoutPlace>().await;
 
-    App::set_window_size((240, 240)).await;
+    AppRunner::set_window_size((240, 240)).await;
 
     inject_touches(
         r"
@@ -171,7 +171,7 @@ pub async fn test_layout() -> Result<()> {
         "|le_s_ct|left|le_ct|center|ri_ct|right|ri_s_ct|bo_s_ct|bottom|bt_ct|center|tp_ct|top|to_s_ct",
     );
 
-    App::set_window_size((400, 400)).await;
+    AppRunner::set_window_size((400, 400)).await;
 
     clear_state();
